==> notebooks/02_transform_raw_data_to_time_series.ipynb <==
import pandas as pd

rides = pd.read_parquet("../data/transformed/rides_2022_01.parquet")

rides.head(20)

rides["pickup_hour"] = rides['pickup_datetime'].dt.floor('H')
rides

app_rides = rides.groupby(["pickup_hour", "pickup_location_id"]).size().reset_index()
app_rides.rename(columns = {0: 'rides'}, inplace = True)
app_rides

from tqdm import tqdm

def add_missing_slots(agg_rides: pd.DataFrame) -> pd.DataFrame:

    location_ids = agg_rides["pickup_location_id"].unique()
    full_range = pd.date_range(
        agg_rides["pickup_hour"].min(),
        agg_rides["pickup_hour"].max(),
        freq="H"
    )
    output = pd.DataFrame()

    for location_id in tqdm(location_ids):
        agg_rides_i = agg_rides.loc[agg_rides.pickup_location_id == location_id, ['pickup_hour', 'rides']]

        agg_rides_i.set_index("pickup_hour", inplace=True)
        agg_rides_i.index = pd.to_datetime(agg_rides_i.index)

        agg_rides_i = agg_rides_i.reindex(full_range, method="ffill")

        agg_rides_i["pickup_location_id"] = location_id
        output = pd.concat([output, agg_rides_i])

    output = output.reset_index().rename(columns={"index": "pickup_hour"})

    return output

agg_rides_all_slots = add_missing_slots(app_rides)

agg_rides_all_slots
