==> notebooks/04_transform_raw_data_into_features_and_targets.ipynb <==
%reload_ext autoreload
%autoreload 2

from src.data import load_raw_data
rides = load_raw_data(year =2022)
rides

rides.isnull().sum()

from src.data import transform_raw_data_into_ts_data

rides = transform_raw_data_into_ts_data(rides)

rides.isnull().sum()

from src.data import transform_ts_data_into_features_and_target

features, target = transform_ts_data_into_features_and_target(
rides,
n_sez_len = 24*28*1,
step_size = 24
)

print(f'{features.shape = }')
print(f'{target.shape = }')

tabular_data = features
tabular_data['target_rides_next_hour'] = target

from src.paths import TRANSFORMED_DATA_DIR
tabular_data.to_parquet(TRANSFORMED_DATA_DIR / 'tabular_data.parquet')
