==> notebooks/06_baseline_model.ipynb <==
import pandas as pd
from src.paths import TRANSFORMED_DATA_DIR

df = pd.read_parquet(TRANSFORMED_DATA_DIR / 'tabular_data.parquet')

df

from datetime import datetime
from src.data_split import train_test_split

X_train, y_train, X_test, y_test = train_test_split(df, cutoff_date =  datetime(2022, 6, 1, 0, 0, 0), target_column_name='target_rides_next_hour')

print(f'{X_train.shape = }')
print(f'{y_train.shape = }')
print(f'{X_test.shape = }')
print(f'{y_test.shape = }')

import numpy as np

class BaselineModelPreviousHour:

    def fit(self, X_train: pd.DataFrame, y_train: pd.Series):
        pass

    def predict(self, X_test: pd.DataFrame) -> np.array:

        return X_test[f'rides_previous_1_hour']

model = BaselineModelPreviousHour()
predictions = model.predict(X_test)
predictions

from sklearn.metrics import mean_absolute_error

test_mae  = mean_absolute_error(y_test, predictions)
print(f'test mae: {test_mae}')

import pandas as pd
import numpy as np

class BaselineModelPreviousHour:

    def fit(self, X_train: pd.DataFrame, y_train: pd.Series):
        pass

    def predict(self, X_test: pd.DataFrame) -> np.array:

        return X_test[f'rides_previous_{7*24}_hour']

model = BaselineModelPreviousHour()
predictions = model.predict(X_test)

predictions = np.nan_to_num(predictions, nan=0)

print(np.isnan(y_test).any())       # Check for NaN in y_test
print(np.isnan(predictions).any()) # Check for NaN in predictions

test_mae  = mean_absolute_error(y_test, predictions)
print(f'test mae: {test_mae}')

class BaselineModelLast4Weeks:

    def fit(self, X_train: pd.DataFrame, y_train: pd.Series):
        pass

    def predict(self, X_test: pd.DataFrame) -> np.array:

        return 0.25 * (X_test[f'rides_previous_{7*24}_hour'] + \
                       X_test[f'rides_previous_{14*24}_hour'] + \
                       X_test[f'rides_previous_{21*24}_hour'] + \
                         X_test[f'rides_previous_{28*24}_hour'])

model = BaselineModelLast4Weeks()
predictions = model.predict(X_test)
predictions = np.nan_to_num(predictions, nan=0)
test_mae  = mean_absolute_error(y_test, predictions)
print(f'test mae: {test_mae}')
