==> notebooks/09_light_gbm_with_feature_engineering.ipynb <==
import pandas as pd
from src.paths import TRANSFORMED_DATA_DIR

df = pd.read_parquet(TRANSFORMED_DATA_DIR / 'tabular_data.parquet')

df

from datetime import datetime
from src.data_split import train_test_split

X_train, y_train, X_test, y_test = train_test_split(df, 
                                                    cutoff_date =  datetime(2022, 6, 1, 0, 0, 0),
                                                    target_column_name='target_rides_next_hour')

print(f'{X_train.shape = }')
print(f'{y_train.shape = }')
print(f'{X_test.shape = }')
print(f'{y_test.shape = }')

def average_rides_past_four_weeks(X: pd.DataFrame) -> pd.DataFrame:
    required_columns = [
        f'rides_previous_{7*24}_hour',
        f'rides_previous_{2*7*24}_hour',
        f'rides_previous_{3*7*24}_hour',
        f'rides_previous_{4*7*24}_hour'
    ]

    # Check if all required columns exist
    missing_columns = [col for col in required_columns if col not in X.columns]
    if missing_columns:
        raise KeyError(f"Missing columns in input DataFrame: {missing_columns}")

    # Calculate the average
    X['average_rides_last_4_weeks'] = (
        X[f'rides_previous_{7*24}_hour'] +
        X[f'rides_previous_{2*7*24}_hour'] +
        X[f'rides_previous_{3*7*24}_hour'] +
        X[f'rides_previous_{4*7*24}_hour']
    ) / 4

    return X


from sklearn.preprocessing import FunctionTransformer

add_feature_average_rides_last_4_weeks = FunctionTransformer(average_rides_past_four_weeks
                                                             , validate=False)


add_feature_average_rides_last_4_weeks.fit_transform(X_train)

from sklearn.base import BaseEstimator, TransformerMixin

class TemporalFeatureTransformer(BaseEstimator, TransformerMixin):

    def fit(self, X, y = None):
        return self

    def transform(self, X, y = None):

        X_ = X.copy()

        X_["hour"] = X_["pickup_hour"].dt.hour
        X_["day_of_week"] = X_["pickup_hour"].dt.dayofweek

        return X_.drop(columns = ["pickup_hour"])

add_temporal_feature = TemporalFeatureTransformer()
add_temporal_feature.fit_transform(X_train)

import lightgbm as lgb

from sklearn.pipeline import make_pipeline

pipeline = make_pipeline(
    add_feature_average_rides_last_4_weeks,
    add_temporal_feature,
    lgb.LGBMRegressor()
)

pipeline.fit(X_train, y_train)

prediction = pipeline.predict(X_test)
from sklearn.metrics import mean_absolute_error

test_mae = mean_absolute_error(y_test, prediction)
print(f'{test_mae}')
