==> notebooks/11_backfill_feature_store.ipynb <==
import os
from dotenv import load_dotenv
from src.paths import PARENT_DIR
from src.data import transform_raw_data_into_ts_data
import hopsworks
import src.config as config
from confluent_kafka import Producer


load_dotenv(PARENT_DIR / ".env")

HOPSWORK_API_KEY = os.environ["HOPSWORKS_API_KEY"]

project = hopsworks.login(
    project = config.HOPSWORKS_PROJECT_NAME,
    api_key_value = HOPSWORK_API_KEY
)
feature_store = project.get_feature_store()
FEATURE_GROUP_NAME = config.FEATURE_GROUP_NAME
FEATURE_GROUP_VERSION = 1
feature_store = project.get_feature_store()
feature_group = feature_store.get_or_create_feature_group(
    name = FEATURE_GROUP_NAME,
    version = FEATURE_GROUP_VERSION,
    description = "Time-series data at hourly frequency",
    primary_key = ["pickup_location_id","pickup_hour"],
    event_time = "pickup_hour"
)

from datetime import datetime
import pandas as pd
from src.data import load_raw_data

from_year = 2022
to_year = datetime.now().year
print(f'Downloading raw data from {from_year} to {to_year}')

rides = pd.DataFrame()
for year in range(from_year, to_year + 1):
    rides_one_year = load_raw_data(year)
    ts_data = transform_raw_data_into_ts_data(rides_one_year)
    feature_group.insert(ts_data, write_options={"wait_for_job": False})
    #rides = pd.concat([rides, rides_one_year])
