==> notebooks/12_feature_pipeline.ipynb <==
import src.config as config

from datetime import datetime, timedelta, timezone

import pandas as pd

current_date = pd.to_datetime(datetime.now(timezone.utc)).floor('H')
print(f'{current_date = }')

fetch_date_to = current_date

fetch_date_from = current_date - timedelta(days = 32)

from src.data import load_raw_data

def fetch_batch_raw_data(from_date: datetime, to_date: datetime) -> pd.DataFrame:
    from_date_ = from_date - timedelta(days = 7*52)
    to_date_ = to_date - timedelta(days = 7*52)

    rides = load_raw_data(year = from_date_.year, months = from_date_.month)
    rides = rides[rides.pickup_datetime >= from_date_]

    rides_2 = load_raw_data(year = to_date_.year, months=to_date_.month)
    rides_2 = rides_2[rides_2.pickup_datetime < to_date_]
    rides = pd.concat([rides, rides_2])

    rides["pickup_datetime"] += timedelta(days=7*52)

    rides.sort_values(by = ["pickup_location_id", "pickup_datetime"], inplace = True)
    return rides

rides = fetch_batch_raw_data(from_date=fetch_date_from, to_date=fetch_date_to)

from src.data import transform_raw_data_into_ts_data
ts_data = transform_raw_data_into_ts_data(rides)

location_id_per_pickup_hour = set(list(ts_data.groupby('pickup_hour')['pickup_location_id'].nunique()))
pickup_hour_per_location_id = set(list(ts_data.groupby('pickup_location_id')['pickup_hour'].nunique()))

assert len(location_id_per_pickup_hour) == 1 and len(pickup_hour_per_location_id) == 1

import hopsworks

project = hopsworks.login(
    project = config.HOPSWORKS_PROJECT_NAME,
    api_key_value = config.HOPSWORKS_API_KEY
)

feature_store = project.get_feature_store()

feature_group = feature_store.get_or_create_feature_group(
    name = config.FEATURE_GROUP_NAME,
    version = config.FEATURE_GROUP_VERSION,
    description = "Time-series data at hourly frequency",
    primary_key = ["pickup_location_id","pickup_hour"],
    event_time = "pickup_hour"
)

ts_data['pickup_location_id'] = ts_data['pickup_location_id'].astype('int64')

print(ts_data.dtypes)

feature_group.insert(ts_data, write_options={"wait_for_job": False})
