==> notebooks/13_model_training_pipeline.ipynb <==
%load_ext autoreload
%autoreload 2

import src.config as config
import pandas as pd

import hopsworks

project = hopsworks.login(
    project = config.HOPSWORKS_PROJECT_NAME,
    api_key_value = config.HOPSWORKS_API_KEY
)

feature_store = project.get_feature_store()

feature_group = feature_store.get_or_create_feature_group(
    name = config.FEATURE_GROUP_NAME,
    version = config.FEATURE_GROUP_VERSION,
)

try:
    # create feature view if it doesn't exist yet
    feature_store.create_feature_view(
        name=config.FEATURE_VIEW_NAME,
        version=config.FEATURE_VIEW_VERSION,
        query=feature_group.select_all()
    )
except:
    print('Feature view already existed. Skip creation.')


# get feature view
feature_view = feature_store.get_feature_view(
    name=config.FEATURE_VIEW_NAME,
    version=config.FEATURE_VIEW_VERSION
)

ts_data = feature_view.training_data(
    description = "Time-series hourly taxi rides"
)

ts_data = ts_data[0]

ts_data

ts_data["pickup_hour"].dtype

ts_data.sort_values(by = ["pickup_location_id", "pickup_hour"], inplace = True)
ts_data

type(ts_data)

ts_data['pickup_hour'] = pd.to_datetime(ts_data['pickup_hour'])

# Filter rows where 'pickup_hour' belongs to 2023 or 2024
filtered_ts_data = ts_data[ts_data['pickup_hour'].dt.year.isin([2023, 2024])]

from src.data import transform_ts_data_into_features_and_target

features, target = transform_ts_data_into_features_and_target(filtered_ts_data,
                                                              n_sez_len = 24*28,
                                                              step_size = 23)

features_and_target = features.copy()
features_and_target['target_rides_next_hour'] = target

print(f'{features_and_target.shape = }')

from datetime import date, timedelta
from pytz  import timezone
import pandas as pd
from src.data_split import train_test_split

cut_off_date = pd.to_datetime(date.today() - timedelta(days = 220*1))

print(f'{cut_off_date = }')

# Convert pickup_hour to UTC if needed
features_and_target['pickup_hour'] = pd.to_datetime(features_and_target['pickup_hour'])

features_and_target['pickup_hour'] = features_and_target['pickup_hour'].dt.tz_convert('UTC')

# Localize cut_off_date to UTC
cut_off_date = cut_off_date.tz_localize('UTC')


X_train, y_train, X_test, y_test = train_test_split(features_and_target, 
                                                    cut_off_date, 
                                                    target_column_name = 'target_rides_next_hour'
                                                    )
print(f'{X_train.shape = }')
print(f'{y_train.shape = }')
print(f'{X_test.shape = }')
print(f'{y_test.shape = }')

import numpy as np
from sklearn.model_selection import KFold, TimeSeriesSplit
from sklearn.pipeline import make_pipeline
from sklearn.metrics import mean_absolute_error
import optuna
from src.model import get_pipeline

def objective(trial: optuna.trial.Trial) -> float:
    # Define the hyperparameter space for the model
    hyperparams = {
        "metric": "mae",
        "verbose": -1,
        "num_leaves": trial.suggest_int("num_leaves", 2, 256),
        "feature_fraction": trial.suggest_float("feature_fraction", 0.2, 1.0),
        "bagging_fraction": trial.suggest_float("bagging_fraction", 0.2, 1.0),
        "min_child_samples": trial.suggest_int("min_child_samples", 3, 100),
    }

    tss = TimeSeriesSplit(n_splits = 2)
    scores = []
    for train_index, val_index in tss.split(X_train):
        X_train_, X_val_ = X_train.iloc[train_index, :], X_train.iloc[val_index, :]
        y_train_, y_val_ = y_train.iloc[train_index], y_train.iloc[val_index]

        pipeline = get_pipeline(**hyperparams)
        pipeline.fit(X_train_, y_train_)

        y_pred = pipeline.predict(X_val_)
        score = mean_absolute_error(y_val_, y_pred)
        scores.append(score)

    return np.array(scores).mean()

study = optuna.create_study(direction = "minimize")
study.optimize(objective, n_trials = 1)

best_params = study.best_trial.params

print(f'{best_params = }')

pipeline = get_pipeline(**best_params)
pipeline.fit(X_train, y_train)

predictions = pipeline.predict(X_test)
test_mae = mean_absolute_error(y_test, predictions)
print(f'Test MAE: {test_mae:.2f}')

import joblib
from src.paths import MODEL_DIR

joblib.dump(pipeline, MODEL_DIR / 'model.pkl')

from hsml.schema import Schema
from hsml.model_schema import ModelSchema

input_schema = Schema(X_train)
output_schema = Schema(y_train)
model_schema = ModelSchema(input_schema = input_schema, output_schema = output_schema)

model_registry = project.get_model_registry()

model = model_registry.sklearn.create_model(
    name = 'taxi_demand_model',
    metrics = {"test_mae": test_mae},
    description = "LightGBM Regressor",
    input_example = X_train.sample(),
    model_schema = model_schema
)
model.save(str(MODEL_DIR / 'model.pkl'))
